requirements.txt
# 핵심 라이브러리
psycopg2-binary==2.9.11
requests==2.31.0
python-dotenv==1.0.1
pyupbit==0.2.34
schedule==1.2.1

fastapi
uvicorn

# 데이터 및 시각화
pandas==2.2.1
numpy==1.26.4
matplotlib==3.8.3
plotly==5.19.0
streamlit==1.32.0

# 유틸리티 및 호환성
urllib3<2.0.0
PyJWT==2.8.0
websockets==12.0
python-dateutil==2.9.0
typing_extensions==4.10.0


# 시스템 종속 라이브러리 (버전 고정 해제하여 호환성 확보)
altgraph
future
six